+++ Changelog.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.4.0
 
 * `Configurate::Proxy` returns its own singleton class if the target does not support creating one.
 * Extract `Configurate::Provider::StringHash` as a new base class for hash based providers.